Add `Data.MessagePack.Option` type for nullable fields.
This behaves like Maybe in every respect, except that it has a different
msgpack representation. Where `Nothing` is encoded as 0 and `Just x` is
encoded as `[1, toObject x]`, `Options` are encoded as `nil` or `toObject
x`. This makes `Option` unnestable. It's still possible to nest Options,
but decoding an encoded `Some None` will result in `None`.

Fixes #17.